feat: rename Environment.tsx -> ClusterEnvironmentDrawer, moved to @Pages/GlobalConfigurations/ClusterAndEnvironments & renamed constants accordingly feat: EnvironemtLabels - way to display error message updated
