+++ src/Pages/GlobalConfigurations/ClustersAndEnvironments/ClusterEnvironmentDrawer/types.ts
@@ -2,7 +2,7 @@ import { TagType } from '@devtron-labs/devtron-fe-common-lib'
 
 import { ClusterNamespacesDTO } from '../clustersAndEnvironments.types'
 
-export interface GlobalEnvironmentProps {
+export interface ClusterEnvironmentDrawerProps {
     environmentName: string
     namespace: string
     id: string
@@ -15,18 +15,18 @@ export interface GlobalEnvironmentProps {
     isVirtual: boolean
 }
 
-export interface GlobalEnvironmentFormProps {
+export interface ClusterEnvironmentDrawerFormProps {
     environmentName: string
     namespace: string
     isProduction: boolean
     description: string
 }
 
-export type GetGlobalEnvironmentUpdatePayloadType = Pick<
-    GlobalEnvironmentProps,
+export type GetClusterEnvironmentUpdatePayloadType = Pick<
+    ClusterEnvironmentDrawerProps,
     'clusterId' | 'id' | 'prometheusEndpoint' | 'isVirtual'
 > &
     Partial<Pick<ClusterNamespacesDTO, 'resourceVersion'>> & {
-        data: GlobalEnvironmentFormProps
+        data: ClusterEnvironmentDrawerFormProps
         namespaceLabels?: TagType[]
     }
