--- src/Pages/GlobalConfigurations/ClustersAndEnvironments/GlobalEnvironment/index.ts
@@ -1,1 +1,0 @@
-export * from './GlobalEnvironment'
